backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
qrcode[pil]>=7.4.2
python-barcode>=0.15.1
pyzbar>=0.1.9
Pillow>=10.0.0
reportlab>=4.2.0
aiofiles>=23.2.1
